microservices-jaeger/requirements.txt
flask==2.3.3
requests==2.31.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
redis==5.0.1
#importlib-metadata==8.4.0  # Add this line to pin a compatible version